test: 테스트 오류 해결
- dev용 profile 적용
- 리팩토링에 따라 테스트 코드 위치 수정
